@@ -6,7 +6,7 @@ import com.couchbase.client.java.Cluster;
 import com.playtika.test.common.spring.DependsOnPostProcessor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.config.BeanFactoryPostProcessor;
-import org.springframework.boot.autoconfigure.AutoConfigureAfter;
+import org.springframework.boot.autoconfigure.AutoConfiguration;
 import org.springframework.boot.autoconfigure.AutoConfigureOrder;
 import org.springframework.boot.autoconfigure.condition.ConditionalOnClass;
 import org.springframework.boot.autoconfigure.condition.ConditionalOnExpression;
@@ -17,11 +17,10 @@ import org.springframework.context.annotation.Configuration;
 import static com.playtika.test.couchbase.CouchbaseProperties.BEAN_NAME_EMBEDDED_COUCHBASE;
 
 @Slf4j
-@Configuration
+@AutoConfiguration(afterName = "org.springframework.boot.autoconfigure.data.couchbase.CouchbaseDataAutoConfiguration")
 @AutoConfigureOrder
 @ConditionalOnExpression("${embedded.containers.enabled:true}")
 @ConditionalOnProperty(name = "embedded.couchbase.enabled", matchIfMissing = true)
-@AutoConfigureAfter(name = "org.springframework.boot.autoconfigure.data.couchbase.CouchbaseDataAutoConfiguration")
 public class EmbeddedCouchbaseDependenciesAutoConfiguration {
 
     @Configuration
