@@ -3,6 +3,7 @@ package com.playtika.test.dynamodb;
 import com.amazonaws.services.dynamodbv2.AmazonDynamoDB;
 import com.playtika.test.common.spring.DependsOnPostProcessor;
 import org.springframework.beans.factory.config.BeanFactoryPostProcessor;
+import org.springframework.boot.autoconfigure.AutoConfiguration;
 import org.springframework.boot.autoconfigure.AutoConfigureOrder;
 import org.springframework.boot.autoconfigure.condition.ConditionalOnClass;
 import org.springframework.boot.autoconfigure.condition.ConditionalOnExpression;
@@ -12,7 +13,7 @@ import org.springframework.context.annotation.Configuration;
 
 import static com.playtika.test.dynamodb.DynamoDBProperties.BEAN_NAME_EMBEDDED_DYNAMODB;
 
-@Configuration
+@AutoConfiguration
 @AutoConfigureOrder
 @ConditionalOnExpression("${embedded.containers.enabled:true}")
 @ConditionalOnClass(AmazonDynamoDB.class)
