@@ -5,17 +5,17 @@ import com.playtika.test.common.utils.PackageInstaller;
 import com.playtika.test.common.utils.YumPackageInstaller;
 import com.playtika.test.kafka.properties.KafkaConfigurationProperties;
 import org.springframework.beans.factory.annotation.Qualifier;
+import org.springframework.boot.autoconfigure.AutoConfiguration;
 import org.springframework.boot.autoconfigure.condition.ConditionalOnBean;
 import org.springframework.boot.autoconfigure.condition.ConditionalOnExpression;
 import org.springframework.boot.autoconfigure.condition.ConditionalOnProperty;
 import org.springframework.boot.context.properties.ConfigurationProperties;
 import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
 import org.testcontainers.containers.GenericContainer;
 
 import static com.playtika.test.kafka.properties.KafkaConfigurationProperties.KAFKA_BEAN_NAME;
 
-@Configuration
+@AutoConfiguration
 @ConditionalOnBean({KafkaConfigurationProperties.class})
 @ConditionalOnExpression("${embedded.containers.enabled:true}")
 @ConditionalOnProperty(value = {"embedded.kafka.enabled"}, havingValue = "true", matchIfMissing = true)
