@@ -6,7 +6,7 @@ import org.neo4j.driver.Driver;
 import org.neo4j.ogm.session.Session;
 import org.neo4j.ogm.session.SessionFactory;
 import org.springframework.beans.factory.config.BeanFactoryPostProcessor;
-import org.springframework.boot.autoconfigure.AutoConfigureAfter;
+import org.springframework.boot.autoconfigure.AutoConfiguration;
 import org.springframework.boot.autoconfigure.AutoConfigureOrder;
 import org.springframework.boot.autoconfigure.condition.ConditionalOnClass;
 import org.springframework.boot.autoconfigure.condition.ConditionalOnExpression;
@@ -17,11 +17,10 @@ import org.springframework.context.annotation.Configuration;
 import static com.playtika.test.neo4j.Neo4jProperties.BEAN_NAME_EMBEDDED_NEO4J;
 
 @Slf4j
-@Configuration
+@AutoConfiguration(afterName = "org.springframework.boot.autoconfigure.data.neo4j.Neo4jDataAutoConfiguration")
 @AutoConfigureOrder
 @ConditionalOnExpression("${embedded.containers.enabled:true}")
 @ConditionalOnProperty(name = "embedded.neo4j.enabled", matchIfMissing = true)
-@AutoConfigureAfter(name = "org.springframework.boot.autoconfigure.data.neo4j.Neo4jDataAutoConfiguration")
 public class EmbeddedNeo4jDependenciesAutoConfiguration {
 
     @Configuration
