@@ -1,16 +1,15 @@
 package com.playtika.test.common.properties;
 
 import com.github.dockerjava.api.model.Capability;
+import jakarta.validation.Valid;
+import jakarta.validation.constraints.NotBlank;
+import jakarta.validation.constraints.NotNull;
 import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 import lombok.With;
 import org.springframework.validation.annotation.Validated;
 import org.testcontainers.containers.BindMode;
-
-import javax.validation.Valid;
-import javax.validation.constraints.NotBlank;
-import javax.validation.constraints.NotNull;
 
 import java.time.Duration;
 import java.util.ArrayList;
