@@ -3,10 +3,9 @@ package com.playtika.test.common.utils;
 import com.playtika.test.common.properties.InstallPackageProperties;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
+import org.springframework.beans.factory.InitializingBean;
 import org.testcontainers.containers.Container;
 import org.testcontainers.containers.GenericContainer;
-
-import javax.annotation.PostConstruct;
 
 /**
  * Instead use ToxiPoxy.
@@ -14,14 +13,18 @@ import javax.annotation.PostConstruct;
 @Slf4j
 @RequiredArgsConstructor
 @Deprecated
-public abstract class PackageInstaller {
+public abstract class PackageInstaller implements InitializingBean {
 
     private final InstallPackageProperties properties;
     private final GenericContainer<?> container;
 
     protected abstract void install(String packageToInstall);
 
-    @PostConstruct
+    @Override
+    public void afterPropertiesSet() {
+        installPackages();
+    }
+
     protected void installPackages() {
         String dockerImageName = container.getDockerImageName();
         String containerName = container.getContainerInfo().getName();
